[vm-benchmarking] move generate_payload and sign_module_txn to helpers
